refactor: move linear data structures to own folder
