+++ Makefile
@@ -1,4 +1,4 @@
 .PHONY: test
 
 test-ds:
-	go test -count=1 -v ./tests/data_structures
+	go test -count=1 -v ./tests/data_structures/linear
