+++ src/data_structures/linear/array.go
