+++ src/data_structures/linear/errors.go
