+++ src/data_structures/linear/linked_list.go
