+++ src/data_structures/linear/stack.go
